Introduce /event endpoint to do the actual work
