+++ vercel.json
@@ -5,6 +5,10 @@
     {
       "source": "/",
       "destination": "/api/index"
+    },
+    {
+      "source": "/event",
+      "destination": "/api/event"
     }
   ]
 }
